fix(network printer): sockets were broking working on threating mode
on threating mode sockets wre broken because each treath try to initializate its own connection to
the remote printer, with this change we delay the connection opening until the actual command that
sends information to the printer. The raw command then opens the connection, send the message and
close the connection at once.