Prepare for Windows deployment
This commit configures the pynsist generator to create the
appropriate Windows service to run pywebdriver using nssm
and changes the config file EOL format to CR+LF, so it
can be edited using notepad.

It also creates a configuration script in the desktop, to
allow the user to easily configure the service, which is
automatically restarted after the editor is closed.