Support Toledo protocol for Baxtran TW scales
This relaxes the parsing of the weight so that other Toledo protocols are permitted, where the weight is returned in grams, without decimal separator.
Also, a unit is allowed (g, lb, oz, kg) to be specified and parsed from the config file.
The previous Mettler_Toledo files are simply renamed as toledo

Fixes https://github.com/akretion/pywebdriver/issues/124

Co-authored-by: hugues de keyzer <github@hugues.info>